# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SchulSaniApp API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CustomRole:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        permissions:
          type: array
          items:
            type: string
        createdBy:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - color
        - permissions
        - createdBy
        - createdAt
    User:
      type: object
      properties:
        id:
          type: string
        iservUsername:
          type: string
        displayName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, sanitaeter_leitung_admin, sanitaeter_leitung, teacher, sanitaeter]
        customRoleId:
          type: string
          nullable: true
        customRole:
          $ref: "#/components/schemas/CustomRole"
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - iservUsername
        - displayName
        - email
        - role
        - isActive
        - createdAt
    MissionResponder:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        displayName:
          type: string
        respondedAt:
          type: string
      required:
        - id
        - userId
        - displayName
        - respondedAt
    Mission:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        location:
          type: string
        patientInitials:
          type: string
        treatmentNotes:
          type: string
        status:
          type: string
          enum: [active, closed]
        createdBy:
          type: string
        createdByName:
          type: string
        createdAt:
          type: string
        closedAt:
          type: string
          nullable: true
        responders:
          type: array
          items:
            $ref: "#/components/schemas/MissionResponder"
      required:
        - id
        - title
        - location
        - patientInitials
        - treatmentNotes
        - status
        - createdBy
        - createdByName
        - createdAt
        - responders
    CreateMissionRequest:
      type: object
      properties:
        title:
          type: string
        location:
          type: string
        patientInitials:
          type: string
        treatmentNotes:
          type: string
      required:
        - title
        - location
        - patientInitials
    UpdateMissionRequest:
      type: object
      properties:
        title:
          type: string
        location:
          type: string
        patientInitials:
          type: string
        treatmentNotes:
          type: string
...
    LoaRequest:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        displayName:
          type: string
        reason:
          type: string
        fromDate:
          type: string
        toDate:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewedBy:
          type: string
          nullable: true
        reviewedByName:
          type: string
          nullable: true
        reviewedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - userId
        - displayName
        - reason
        - fromDate
        - toDate
        - status
        - createdAt
    CreateLoaRequest:
      type: object
      properties:
        reason:
          type: string
        fromDate:
          type: string
        toDate:
          type: string
      required:
        - reason
        - fromDate
        - toDate
    UpdateLoaRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
      required:
        - status
    NewsPost:
      type: object
      properties:
        id:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - authorId
        - authorName
        - title
        - content
        - isPinned
        - createdAt
    CreateNewsRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
      required:
        - title
        - content
    UpdateNewsRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
    UpdateUserRequest:
      type: object
      properties:
        role:
          type: string
          enum: [admin, sanitaeter_leitung_admin, sanitaeter_leitung, teacher, sanitaeter]
        isActive:
          type: boolean
        customRoleId:
          type: string
          nullable: true
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
security:
  - bearerAuth: []
paths:
...
  /missions:
    get:
      operationId: getMissions
      tags: [missions]
      summary: List missions
      responses:
        "200":
          description: List of missions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mission"
    post:
      operationId: createMission
      tags: [missions]
      summary: Create mission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMissionRequest"
      responses:
        "201":
          description: Mission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
  /missions/{id}:
    get:
      operationId: getMission
      tags: [missions]
      summary: Get mission by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
    patch:
      operationId: updateMission
      tags: [missions]
      summary: Update mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMissionRequest"
      responses:
        "200":
          description: Mission updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
  /missions/{id}/close:
    post:
      operationId: closeMission
      tags: [missions]
      summary: Close a mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mission closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
  /missions/{id}/respond:
    post:
      operationId: respondToMission
      tags: [missions]
      summary: Respond to a mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
...
  /loa:
    get:
      operationId: getLoaRequests
      tags: [loa]
      summary: List LOA requests
      responses:
        "200":
          description: List of LOA requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoaRequest"
    post:
      operationId: createLoaRequest
      tags: [loa]
      summary: Create LOA request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoaRequest"
      responses:
        "201":
          description: LOA request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoaRequest"
  /loa/{id}:
    patch:
      operationId: updateLoaRequest
      tags: [loa]
      summary: Approve or reject LOA request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLoaRequest"
      responses:
        "200":
          description: LOA updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoaRequest"
  /news:
    get:
      operationId: getNews
      tags: [news]
      summary: List news posts
      responses:
        "200":
          description: List of news posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsPost"
    post:
      operationId: createNews
      tags: [news]
      summary: Create news post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsRequest"
      responses:
        "201":
          description: News post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsPost"
  /news/{id}:
    patch:
      operationId: updateNews
      tags: [news]
      summary: Update news post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNewsRequest"
      responses:
        "200":
          description: News updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsPost"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete news post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user role/status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...